+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.3.1
 * `NEW` setting `hover.previewFields`: limit how many fields are shown in table hover
 * `NEW` fully support `---@type Object[]`
+* `CHG` resolve infer of `string|table`
 
 ## 1.3.0
 `2020-12-1`